reduce size

+++ DQM/DTOfflineAnalysis/src/DTHitObject.h
@@ -4,8 +4,8 @@
 /** \class DTHitObject
  *  No description available.
  *
- *  $Date: 2009/07/16 14:47:09 $
- *  $Revision: 1.3 $
+ *  $Date: 2010/06/07 17:08:49 $
+ *  $Revision: 1.4 $
  *  \author G. Cerminara - INFN Torino
  */
 
@@ -37,35 +37,35 @@ public:
 //   int wheel;
 //   int station;
 //   int sector;
-  int sl;
-  int layer;
-  int wire;
+  short sl;
+  short layer;
+  short wire;
 
   // position in local RF
-  double X;
-  double Y;
-  double Z;
+  float X;
+  float Y;
+  float Z;
 
   // residual on distance from the wire
-  double resDist;
+  float resDist;
   // residual on distance from the wire - step 2
-  double resDistS1;
+  float resDistS1;
   // residual on distance from the wire - step 1
-  double resDistS2;
+  float resDistS2;
   // residual on position
-  double resPos;
+  float resPos;
   // distance from the wire
-  double distFromWire;
+  float distFromWire;
   // sigma on position
-  double sigmaPos;
+  float sigmaPos;
   // the angle
-  double angle;
+  float angle;
   // flag to label noisy channels
   bool isNoisyCell;
   // digi time (ns)
-  double digiTime;
+  float digiTime;
   // t0 from pulses
-  double t0pulses;
+  float t0pulses;
   
   ClassDef(DTHitObject,1)
 };
